Update to use longs instead of ints

+++ src/main/java/io/github/bluelhf/permutations/GenerationData.java
@@ -8,13 +8,13 @@ import java.util.List;
 import java.util.concurrent.CompletableFuture;
 
 public class GenerationData {
-    private int max;
-    private int threadSize;
-    private int threadCount;
+    private long max;
+    private long threadSize;
+    private long threadCount;
     private long timeStarted;
     private @Nullable Generator generator;
     private List<CompletableFuture<Void>> threads = new ArrayList<>();
-    private LinkedHashMap<Integer, Double> threadProgress = new LinkedHashMap<>();
+    private LinkedHashMap<Long, Double> threadProgress = new LinkedHashMap<>();
 
     public GenerationData() {
         this.max = 0;
@@ -24,8 +24,8 @@ public class GenerationData {
         this.generator = null;
     }
 
-    public GenerationData(Generator generator, List<CompletableFuture<Void>> threads, int max, int threadSize, int threadCount, long timeStarted) {
-        for(int i = 0; i < threads.size(); i++) threadProgress.put(i, 0D);
+    public GenerationData(Generator generator, List<CompletableFuture<Void>> threads, long max, long threadSize, long threadCount, long timeStarted) {
+        for(long i = 0; i < threads.size(); i++) threadProgress.put(i, 0D);
         this.threads.addAll(threads);
         this.max = max;
         this.threadSize = threadSize;
@@ -43,10 +43,10 @@ public class GenerationData {
         return this;
     }
 
-    public double getProgress(int threadId) {
+    public double getProgress(long threadId) {
         return threadProgress.get(threadId);
     }
-    public GenerationData setProgress(int threadId, double progress) {
+    public GenerationData setProgress(long threadId, double progress) {
         this.threadProgress.put(threadId, progress);
         return this;
     }
@@ -58,27 +58,27 @@ public class GenerationData {
     public GenerationData setThreads(List<CompletableFuture<Void>> newThreads) {
         this.threads.clear();
         this.threads.addAll(newThreads);
-        for(int i = 0; i < this.threads.size(); i++) threadProgress.put(i, 0D);
+        for(long i = 0; i < this.threads.size(); i++) threadProgress.put(i, 0D);
         return this;
     }
 
-    public GenerationData setThreadProgress(LinkedHashMap<Integer, Double> newThreadProgress) {
+    public GenerationData setThreadProgress(LinkedHashMap<Long, Double> newThreadProgress) {
         this.threadProgress.clear();
         this.threadProgress.putAll(newThreadProgress);
         return this;
     }
 
-    public GenerationData setMax(int newMax) {
+    public GenerationData setMax(long newMax) {
         this.max = newMax;
         return this;
     }
 
-    public GenerationData setThreadSize(int newThreadSize) {
+    public GenerationData setThreadSize(long newThreadSize) {
         this.threadSize = newThreadSize;
         return this;
     }
 
-    public GenerationData setThreadCount(int newThreadCount) {
+    public GenerationData setThreadCount(long newThreadCount) {
         this.threadCount = newThreadCount;
         return this;
     }
@@ -88,15 +88,15 @@ public class GenerationData {
         return this;
     }
 
-    public int getThreadCount() {
+    public long getThreadCount() {
         return threadCount;
     }
 
-    public int getThreadSize() {
+    public long getThreadSize() {
         return threadSize;
     }
 
-    public int getMax() {
+    public long getMax() {
         return max;
     }
